Move manager to managers.py
