+++ sponsors/managers.py
@@ -1,3 +1,5 @@
+from django.db.models import Count
+from ordered_model.models import OrderedModelManager
 from django.db.models import Q, Subquery
 from django.db.models.query import QuerySet
 
@@ -25,3 +27,21 @@ class SponsorContactQuerySet(QuerySet):
         if not contact:
             raise self.model.DoesNotExist()
         return contact
+
+
+class SponsorshipBenefitManager(OrderedModelManager):
+    def with_conflicts(self):
+        return self.exclude(conflicts__isnull=True)
+
+    def without_conflicts(self):
+        return self.filter(conflicts__isnull=True)
+
+    def add_ons(self):
+        return self.annotate(num_packages=Count("packages")).filter(num_packages=0)
+
+    def with_packages(self):
+        return (
+            self.annotate(num_packages=Count("packages"))
+            .exclude(num_packages=0)
+            .order_by("-num_packages")
+        )
